+++ src/KABubbleWindowView.m
@@ -5,11 +5,11 @@
 #import "KABubbleWindowView.h"
 
 // info needs to be a KABubbleWindowView with rgb+alpha set.
-void KABubbleShadeInterpolate( void *info, float const *inData, float *outData ) {
+void KABubbleShadeInterpolate( void *info, CGFloat const *inData, CGFloat *outData ) {
 	// These 2 will always return an array of 4 floats
-	const float *dark = [(KABubbleWindowView *)info darkColorFloat];
-	const float *light = [(KABubbleWindowView *)info lightColorFloat];
-	float a = inData[0];
+	const CGFloat *dark = [(KABubbleWindowView *)info darkColorFloat];
+	const CGFloat *light = [(KABubbleWindowView *)info lightColorFloat];
+	CGFloat a = inData[0];
 	int i = 0;
 
 	for( i = 0; i < 4; i++ )
@@ -163,7 +163,7 @@ void KABubbleShadeInterpolate( void *info, float const *inData, float *outData )
 {
 	return _darkColor;
 }
-- (const float *) darkColorFloat
+- (const CGFloat *) darkColorFloat
 {
 	return _darkColorFloat;
 }
@@ -185,7 +185,7 @@ void KABubbleShadeInterpolate( void *info, float const *inData, float *outData )
 {
 	return _lightColor;
 }
-- (const float *) lightColorFloat
+- (const CGFloat *) lightColorFloat
 {
 	return _lightColorFloat;
 }
